move WarningBox from forms directory
