+++ app/components/widgets/WarningBox.js
