+++ app/components/widgets/WarningBox.scss
@@ -15,7 +15,7 @@
     .headerIcon {
       width: 20px;
       height: 20px;
-      background-image: url(../../../assets/images/attention.inline.svg);
+      background-image: url(../../assets/images/attention.inline.svg);
       background-size: contain;
       background-position: center center;
       background-repeat: no-repeat;
